refactor: Further decouple API hooks from framework
